Type selection object in editOnInput.
Summary:
Recently I merged D20582792— it was a quick fix to a a huge problem that was causing thousands of exceptions for our users.

It could've been prevented with better typing, but wasn't caught by flow because the type of the selection object was assumed to be null.

Here I implement the type of the selection object, and in the process catch other elements that could be null and add safeguards to make our code more reliable. This are places where the previous implementation would've thrown an exception.

Reviewed By: kedromelon

Differential Revision: D20636236

fbshipit-source-id: 6e27040120256e3885bfdd2d9cbedf5e57988062