Make sure that the placeholder component checks all of the props in its `shouldComponentUpdate`
Summary:
I found this `shouldComponentUpdate` method that wasn't checking all of the props (eg. `accessibilityID`). This new strategy is to check the focus state first, //then// to shallow compare the rest of the props.

I do this, realizing that this is probably overkill to begin with, but it's now more correct!

Reviewed By: mrkev

Differential Revision: D23912258

fbshipit-source-id: f14fc75d0e9edca3da4918330262d479884a59ab