Strongly type ContentState.getEntityMap
Summary:
In the quest to deprecate the old DraftEntity API, we want to leverage flow as much as possible, so this any here is pretty undesirable.

I removed it and it revealed an instance where we want to use new API, so I changed that too.

On the tests— this is just mocking behaviour, so leaving the fix for later.

Reviewed By: shalabhvyas, AjkP

Differential Revision: D24038382

fbshipit-source-id: 0d63074389507a5e6bd15b7752d8a544da150b15