Add helper superclass type to ContentState
Summary:
Here's a fun one: `ContentStateRecord` is of type any, so subclassing `ContentState` from it means our new class can have essentially any method.

If its a method defined in the `ContentState`, its type will respected. If it's not, we'll just assume its there with in/out any. I'm assuming this applies to fields too.

I noticed this because I want to remove a method, but alas, I did and instead of pointing out where the uses are, flow returned the usually-great-but-today-stressful "No Errors!" message. Lovely.

In this diff I'm manually declaring a type for the superclass, so when I remove the method its pointed out by flow. It's full of anys, but it's still better than what we had before.

How do I know this? Because it outright already found bugs. There's callsites with the wrong number of argumetns, there's people calling this methods that don't exist, *sigh*. These are instant exceptions at best, deeply nested, obscure bugs otherwise.

Reviewed By: AjkP

Differential Revision: D24068574

fbshipit-source-id: 32b870e360fed003fef1d0b9800cdf15e5d90dd0