Migrate ContentState to safer setter API 3/3
Summary:
Currently, to set the fields of some `ContentState`, users use the immutable API. That is, they call `.set('some_string', any_value)`.

Having specific, type-safe setters is 100x safer. This is the last diff and gets rid of the last few instances where we're calling `.set()` directly.

Now, here is the interesting bit— we had calls to `ContentState.set('entityMap', ...)` in the code. This is weird for two reasons.

1. Under the current architecture, it's a no-op.
2. This is what we're trying to migrate to, so under the new architecture, it'll _do something_— hopefully the right thing?

Moreover, 2/3 files doing this weren't even random product code— they are in the Draft.js "core". So what gives? What's the right thing to do here? How is nothing broken when these seemingly important lines of code are acutally not doing anything? Why are these no-ops here? WTF is going on?

Blame to the rescue.

## The story

In essence, the migration away from a global Draft Entity was attempted before, in: https://github.com/facebook/draft-js/commit/5d5f1b4b89c509d17697965ce9a0d596ed220c43

That didn’t work out, so things were reverted back to the way they are now. Nonetheless, these `.set('entityMap', ...)` weren’t removed. The only file that wasn’t part of that process is `EditorCorePasteOverride.js`, which seems to just have been copied over from `editOnPaste.js` in D14626028 (https://github.com/facebook/draft-js/commit/d09ef3e416beb766b5a227ea84e8d735ff11823b), `.set('entityMap', ...)` included.

## This change

Having this was necessary in a world without a global DraftEntity because it updates the entity map in our content state. When DraftEntity is global, it’s mutable, so some other function can (and does) set the entities without one needing to merge maps and/or set the new map to the `ContentState`.

On this diff I do do it though, because effectively setting it is the same, but more correct. And it means when we migrate out of a global `DraftEntity` we don’t have to come back and change this code.

Reviewed By: AjkP

Differential Revision: D24698388

fbshipit-source-id: ad530f112e2f5e4db6657e44136aeb0f0fc1635e