Improve handling of internal pastes
Summary:
Draft.js's default handling of pasted content (`editOnPaste.js`) includes a special case for when it's an internal paste (content copy-pasted from within the same editor). It also supports specifying a custom paste-handling function, which can override the former.

There are legitimate use cases for wanting to specify custom paste handling which only applies to non-internal pastes (or equivalently, for that custom paste handler to receive a flag for whether this is an internal paste, which is what this commit is introducing), due to that default internal paste handling being preferable when possible.

For example, an editor which ordinarily has a custom paste handler (e.g. for formatting the pasted content) might choose to skip that handling for internal pastes.

Reviewed By: mrkev

Differential Revision: D25558780

fbshipit-source-id: bcbf3cd8c12034dedce62930789723bf19b8da93